IGVF-60-report-view
Rename session-storage.js to browser-storage.js. Add a localStorage React custom hook.
